b4: Include git notes when generating patches
By default, 'git show' will include git notes in its output. However,
this is not the case if '--format', or related flags, are given.

Include the '--notes' flag to include git notes in the generated patch.

Also, git-filter-repo does not use git rebase when changing history. So
the default 'notes rewrite' flow will not work. The notes will still be
available, but they will not be associated with the newly created
objects. Fortunately, git-filter-repo records the list of changed objects
after every operation.

Add a callback to copy the notes to the newly created objects. Use the
'--force' flag to avoid benign warnings related to existing notes, etc.
Notes will be copied from an old object to a new object which should not
have any existing notes. Or notes will be copied from an old object to
itself (unchanged), so overwriting its notes is not an issue.

Based on the discussion here:
https://github.com/newren/git-filter-repo/issues/22

This should have no effect on non-users of git notes.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>