b4: Check excludes when creating extracc list for patch summaries
The per-patch summary does not check for excluded addresses when forming
the extracc list.

Example with email-exclude config and '--not-me-too' flag:

[b4]
        email-exclude = stable@vger.kernel.org

$ b4 send --not-me-too
Converted the branch to 1 messages
---
To: Yazen Ghannam <Yazen.Ghannam@amd.com>
Cc: linux-kernel@vger.kernel.org
---
  [PATCH] Example
    +Cc: Yazen Ghannam <yazen.ghannam@amd.com>
         stable@vger.kernel.org
---

The final list is correct, but this is not visible to users without
checking debug logs.

Apply the excludes check to extracc before printing so users will be
aware that their exclude rules are applied.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>